updated login.test.js ->
Removed: code that saves both token and profile to localStorage after login() is runned. This proved to trick the test to always be equal what I set to expect. Whether login() is runned or not.

Removed: JSON.stringify() of TEST_PROFILE object. This to be able to access the accessToken from mockStorage

Added: JSON.parse() of localStorage receiving accessToken, to prove it is the same as TEST_TOKEN.